refactor(daemon): split tools-connectors-cli god-file into a capability-barrel module
`apps/daemon/src/tools-connectors-cli.ts` had grown to 2,830 lines fusing two
subsystems — the connectors tool CLI and the design-system package-audit engine
— plus shared plumbing, in one flat file. Split it into a `tools-connectors-cli/`
capability-barrel module following the `design-systems/` reference: a `core/`
foundation kernel plus concern subdirectories (`intake/`, `audit/`, `evidence/`),
each with a barrel, a root `index.ts` barrel, and a module `README.md`.

Layout is an acyclic layered graph — `core/` imports no sibling; `intake/ -> core`;
`audit/ -> core`; `evidence/ -> audit, core`; `commands.ts` (root dispatch) -> all
subdir barrels. Cross-subdir imports route only through sibling barrels; external
code imports only the root barrel. The public export surface is unchanged and the
four importers keep importing `./tools-connectors-cli/index.js`.

Function bodies are byte-identical to the original (the concern partition was
proven equal to the original file body), with one deliberate cycle-breaking
relocation: `packageBuildAssetTarget` moved into `core/github-paths.ts` so
`audit` no longer imports `evidence`. Every file carries a `@module` docblock and
every export is JSDoc'd.

The machine-enforced barrel guard (`scripts/check-barrel-imports.ts`) is not on
`main` yet, so this module is not registered in `CAPABILITY_BARREL_DOMAINS`; the
acyclic layering is maintained by convention until that infra lands.

No behavior change; public API identical.
