+++ apps/daemon/src/cli.ts
@@ -6,7 +6,7 @@ import { runDaemonCliStartup, startDaemonRuntime } from './daemon-startup.js';
 import { runLiveArtifactsMcpServer } from './mcp-live-artifacts-server.js';
 import { runArtifactsCli } from './artifacts-cli.js';
 import { runProjectHandoff } from './handoff-cli.js';
-import { runConnectorsToolCli } from './tools-connectors-cli.js';
+import { runConnectorsToolCli } from './tools-connectors-cli/index.js';
 import { runDesignSystemsToolCli } from './tools-design-systems-cli.js';
 import { DESIGN_SYSTEMS_USAGE, isDesignSystemsHelpArg } from './design-systems-cli-help.js';
 import { BRAND_USAGE, isBrandHelpArg } from './brands-cli-help.js';
