+++ apps/daemon/src/tools-connectors-cli/audit/index.ts
@@ -1,0 +1,6 @@
+/** @module audit/index
+ * The design-system package-audit engine: scores a candidate package's design
+ * evidence against the reuse / preview-manifest / token-binding rules and returns
+ * structured issues (`DesignSystemPackageAudit`). Depends only on core.
+ */
+export * from './audit.js';
