+++ apps/daemon/src/tools-connectors-cli/evidence/index.ts
@@ -1,0 +1,6 @@
+/** @module evidence/index
+ * Evidence artifact writing: serializes collected design evidence to JSON and
+ * renders the markdown context plus materialized package assets. Depends on audit
+ * (component-name helpers) and core.
+ */
+export * from './evidence-write.js';
